Separate bin from lib
The CLI interface is named jan because it creates jan files ^-^

I know unwrap is a little sketchy, but I can't decide how else to handle the error without potentially causing issues in scripts.
